fix: use dumpTextRow to print
