+++ util/sqlexec/BUILD.bazel
@@ -17,10 +17,8 @@ go_library(
         "//types",
         "//util/chunk",
         "//util/hack",
-        "//util/logutil",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_kvproto//pkg/kvrpcpb",
-        "@org_uber_go_zap//:zap",
     ],
 )
